fix: Widen Fernet key index from i8 to u32
Python Keystone's key index is an unbounded int; the Rust i8 type
capped it at 127, silently mishandling repositories with higher
indices (e.g. from long rotation history or migration from Python).

Add tests covering indices past the old i8 cap for both
FilesystemKeySource::load and KeyRepository::load_keys/rotate.

+++ crates/credential-driver-sql/src/error.rs
@@ -31,7 +31,7 @@ pub enum CredentialFernetError {
 
     /// A key file's contents could not be parsed as a Fernet key.
     #[error("key file `{0}` is not a valid Fernet key")]
-    InvalidKey(i8),
+    InvalidKey(u32),
 
     /// No usable key files were found in the repository.
     #[error("no Fernet keys found in the credential key repository")]
@@ -45,7 +45,7 @@ pub enum CredentialFernetError {
     )]
     NullKeyDetected,
 
-    /// Fernet index arithmetic would overflow the `i8` file-naming scheme.
+    /// Fernet index arithmetic would overflow the `u32` file-naming scheme.
     #[error("key rotation index overflow")]
     IndexOverflow,
